backend/services/content/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.5.0
pydantic-settings==2.1.0
httpx==0.27.0
python-multipart==0.0.9
boto3==1.34.0
Pillow==10.2.0
elasticsearch==8.12.0
pandas==2.2.0
openpyxl==3.1.2
pymysql==1.1.1
google-generativeai
redis
six
hangulize==0.0.9
google-api-python-client==2.118.0
sqlalchemy==2.0.25

opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
slowapi==0.1.9

# Lambda support
mangum==0.17.0
